+++ lib/wire/const.js
@@ -1,4 +1,245 @@
-exports.DEFAULT_ENCODING = 'UTF8';
+/***************************************
+ *
+ *   Constantes
+ *
+ ***************************************/
+
+const defaultOptions = {
+    DEFAULT_HOST : '127.0.0.1',
+    DEFAULT_PORT : 3050,
+    DEFAULT_USER : 'SYSDBA',
+    DEFAULT_PASSWORD : 'masterkey',
+    DEFAULT_LOWERCASE_KEYS : false,
+    DEFAULT_PAGE_SIZE : 4096,
+    DEFAULT_SVC_NAME : 'service_mgr',
+    DEFAULT_ENCODING : 'UTF8',
+    DEFAULT_FETCHSIZE : 200,
+};
+
+const buffer= {
+    MAX_BUFFER_SIZE : 8192,
+};
+
+const int = {
+    MAX_INT : Math.pow(2, 31) - 1,
+    MIN_INT : -Math.pow(2, 31),
+};
+
+const op = {
+    op_void                   : 0,  // Packet has been voided
+    op_connect                : 1,  // Connect to remote server
+    op_exit                   : 2,  // Remote end has exitted
+    op_accept                 : 3,  // Server accepts connection
+    op_reject                 : 4,  // Server rejects connection
+    op_disconnect             : 6,  // Connect is going away
+    op_response               : 9,  // Generic response block
+
+    // Full context server operations
+
+    op_attach                 : 19, // Attach database
+    op_create                 : 20, // Create database
+    op_detach                 : 21, // Detach database
+    op_compile                : 22, // Request based operations
+    op_start                  : 23,
+    op_start_and_send         : 24,
+    op_send                   : 25,
+    op_receive                : 26,
+    op_unwind                 : 27, // apparently unused, see protocol.cpp's case op_unwind
+    op_release                : 28,
+
+    op_transaction            : 29, // Transaction operations
+    op_commit                 : 30,
+    op_rollback               : 31,
+    op_prepare                : 32,
+    op_reconnect              : 33,
+
+    op_create_blob            : 34, // Blob operations
+    op_open_blob              : 35,
+    op_get_segment            : 36,
+    op_put_segment            : 37,
+    op_cancel_blob            : 38,
+    op_close_blob             : 39,
+
+    op_info_database          : 40, // Information services
+    op_info_request           : 41,
+    op_info_transaction       : 42,
+    op_info_blob              : 43,
+
+    op_batch_segments         : 44, // Put a bunch of blob segments
+
+    op_que_events             : 48, // Que event notification request
+    op_cancel_events          : 49, // Cancel event notification request
+    op_commit_retaining       : 50, // Commit retaining (what else)
+    op_prepare2               : 51, // Message form of prepare
+    op_event                  : 52, // Completed event request (asynchronous)
+    op_connect_request        : 53, // Request to establish connection
+    op_aux_connect            : 54, // Establish auxiliary connection
+    op_ddl                    : 55, // DDL call
+    op_open_blob2             : 56,
+    op_create_blob2           : 57,
+    op_get_slice              : 58,
+    op_put_slice              : 59,
+    op_slice                  : 60, // Successful response to op_get_slice
+    op_seek_blob              : 61, // Blob seek operation
+
+// DSQL operations
+
+    op_allocate_statement     : 62, // allocate a statment handle
+    op_execute                : 63, // execute a prepared statement
+    op_exec_immediate         : 64, // execute a statement
+    op_fetch                  : 65, // fetch a record
+    op_fetch_response         : 66, // response for record fetch
+    op_free_statement         : 67, // free a statement
+    op_prepare_statement      : 68, // prepare a statement
+    op_set_cursor             : 69, // set a cursor name
+    op_info_sql               : 70,
+
+    op_dummy                  : 71, // dummy packet to detect loss of client
+    op_response_piggyback     : 72, // response block for piggybacked messages
+    op_start_and_receive      : 73,
+    op_start_send_and_receive : 74,
+    op_exec_immediate2        : 75, // execute an immediate statement with msgs
+    op_execute2               : 76, // execute a statement with msgs
+    op_insert                 : 77,
+    op_sql_response           : 78, // response from execute, exec immed, insert
+    op_transact               : 79,
+    op_transact_response      : 80,
+    op_drop_database          : 81,
+    op_service_attach         : 82,
+    op_service_detach         : 83,
+    op_service_info           : 84,
+    op_service_start          : 85,
+    op_rollback_retaining     : 86,
+    op_partial                : 89, // packet is not complete - delay processing
+    op_trusted_auth           : 90,
+    op_cancel                 : 91,
+    op_cont_auth              : 92,
+    op_ping                   : 93,
+    op_accept_data            : 94, // Server accepts connection and returns some data to client
+    op_abort_aux_connection   : 95, // Async operation - stop waiting for async connection to arrive
+    op_crypt                  : 96,
+    op_crypt_key_callback     : 97,
+    op_cond_accept            : 98, // Server accepts connection, returns some data to client
+                                    // and asks client to continue authentication before attach call
+};
+
+const dsql = {
+    DSQL_close : 1,
+    DSQL_drop : 2,
+    DSQL_unprepare : 4, // >: 2.5
+};
+
+/***********************/
+/*   ISC Error Codes   */
+/***********************/
+const iscError = {
+    isc_sqlerr: 335544436,
+    isc_arg_end : 0,  // end of argument list
+    isc_arg_gds : 1,  // generic DSRI status value
+    isc_arg_string : 2,  // string argument
+    isc_arg_cstring : 3,  // count & string argument
+    isc_arg_number : 4,  // numeric argument (long)
+    isc_arg_interpreted : 5,  // interpreted status code (string)
+    isc_arg_unix : 7,  // UNIX error code
+    isc_arg_next_mach : 15, // NeXT/Mach error code
+    isc_arg_win32 : 17, // Win32 error code
+    isc_arg_warning : 18, // warning argument
+    isc_arg_sql_state : 19, // SQLSTATE
+};
+
+const connect = {
+    CONNECT_VERSION2 : 2,
+    CONNECT_VERSION3 : 3,
+    ARCHITECTURE_GENERIC : 1,
+};
+
+/*******************/
+/*    Protocols    */
+/*******************/
+const FB_PROTOCOL_FLAG = 0x8000;
+const protocol = {
+    // Protocol 10 includes support for warnings and removes the requirement for
+    // encoding and decoding status codes
+    PROTOCOL_VERSION10  : 10,
+
+    // Since protocol 11 we must be separated from Borland Interbase.
+    // Therefore always set highmost bit in protocol version to 1.
+    // For unsigned protocol version this does not break version's compare.
+    FB_PROTOCOL_FLAG    : FB_PROTOCOL_FLAG,
+    FB_PROTOCOL_MASK    : ~FB_PROTOCOL_FLAG & 0xFFFF,
+
+    // Protocol 11 has support for user authentication related
+    // operations (op_update_account_info, op_authenticate_user and
+    // op_trusted_auth). When specific operation is not supported,
+    // we say "sorry".
+    PROTOCOL_VERSION11  : (FB_PROTOCOL_FLAG | 11),
+
+    // Protocol 12 has support for asynchronous call op_cancel.
+    // Currently implemented asynchronously only for TCP/IP.
+    PROTOCOL_VERSION12  : (FB_PROTOCOL_FLAG | 12),
+
+    // Protocol 13 has support for authentication plugins (op_cont_auth).
+    PROTOCOL_VERSION13  : (FB_PROTOCOL_FLAG | 13),
+};
+
+// Protocols types (accept_type)
+const acceptType = {
+    ptype_rpc : 2, 			 // Simple remote procedure call
+    ptype_batch_send : 3,    // Batch sends, no asynchrony
+    ptype_out_of_band : 4,   // Batch sends w/ out of band notification
+    ptype_lazy_send : 5,     // Deferred packets delivery;
+    ptype_mask : 0xFF,       // Mask - up to 255 types of protocol
+    pflag_compress : 0x100  // Turn on compression if possible
+};
+
+const SUPPORTED_PROTOCOL = [
+    [protocol.PROTOCOL_VERSION10, connect.ARCHITECTURE_GENERIC, acceptType.ptype_rpc, acceptType.ptype_batch_send, 1],
+    [protocol.PROTOCOL_VERSION11, connect.ARCHITECTURE_GENERIC, acceptType.ptype_lazy_send, acceptType.ptype_lazy_send, 2],
+    [protocol.PROTOCOL_VERSION12, connect.ARCHITECTURE_GENERIC, acceptType.ptype_lazy_send, acceptType.ptype_lazy_send, 3],
+    [protocol.PROTOCOL_VERSION13, connect.ARCHITECTURE_GENERIC, acceptType.ptype_lazy_send, acceptType.ptype_lazy_send, 4],
+];
+
+const authPlugin = {
+    AUTH_PLUGIN_LEGACY : 'Legacy_Auth',
+    AUTH_PLUGIN_SRP : 'Srp',
+    // AUTH_PLUGIN_SRP256 : 'Srp256',
+};
+
+const authOptions = {
+    // AUTH_PLUGIN_LIST : [authPlugin.AUTH_PLUGIN_SRP256, authPlugin.AUTH_PLUGIN_SRP, authPlugin.AUTH_PLUGIN_LEGACY],
+    AUTH_PLUGIN_LIST : [authPlugin.AUTH_PLUGIN_SRP, authPlugin.AUTH_PLUGIN_LEGACY],
+    // AUTH_PLUGIN_SRP_LIST : [authPlugin.AUTH_PLUGIN_SRP256, authPlugin.AUTH_PLUGIN_SRP],
+    AUTH_PLUGIN_SRP_LIST : [authPlugin.AUTH_PLUGIN_SRP],
+    LEGACY_AUTH_SALT : '9z',
+    WIRE_CRYPT_DISABLE : 0,
+    WIRE_CRYPT_ENABLE : 1,
+};
+
+/*******************/
+/*    SQL Type     */
+/*******************/
+const sqlType = {
+    SQL_TEXT : 452, // Array of char
+    SQL_VARYING : 448,
+    SQL_SHORT : 500,
+    SQL_LONG : 496,
+    SQL_FLOAT : 482,
+    SQL_DOUBLE : 480,
+    SQL_D_FLOAT : 530,
+    SQL_TIMESTAMP : 510,
+    SQL_BLOB : 520,
+    SQL_ARRAY : 540,
+    SQL_QUAD : 550,
+    SQL_TYPE_TIME : 560,
+    SQL_TYPE_DATE : 570,
+    SQL_INT64 : 580,
+    SQL_BOOLEAN : 32764, // >: 3.0
+    SQL_NULL : 32766, // >= 2.5
+};
+
+const blobType = {
+    isc_blob_text : 1,
+};
 
 /*******************/
 /* Blr definitions */
@@ -133,22 +374,22 @@ const dpb = {
     isc_dpb_session_time_zone 		: 91,
 };
 
-// const cnct = {
-//     CNCT_user : 1, // User name
-//     CNCT_passwd : 2,
-//     // CNCT_ppo : 3, // Apollo person, project, organization. OBSOLETE.
-//     CNCT_host : 4,
-//     CNCT_group : 5, // Effective Unix group id
-//     CNCT_user_verification : 6, // Attach/create using this connection will use user verification
-//     CNCT_specific_data : 7, // Some data, needed for user verification on server
-//     CNCT_plugin_name : 8, // Name of plugin, which generated that data
-//     CNCT_login : 9, // Same data as isc_dpb_user_name
-//     CNCT_plugin_list : 10, // List of plugins, available on client
-//     CNCT_client_crypt : 11, // Client encyption level (DISABLED/ENABLED/REQUIRED)
-//     WIRE_CRYPT_DISABLED : 0,
-//     WIRE_CRYPT_ENABLED : 1,
-//     WIRE_CRYPT_REQUIRED : 2,
-// };
+const cnct = {
+    CNCT_user : 1, // User name
+    CNCT_passwd : 2,
+    // CNCT_ppo : 3, // Apollo person, project, organization. OBSOLETE.
+    CNCT_host : 4,
+    CNCT_group : 5, // Effective Unix group id
+    CNCT_user_verification : 6, // Attach/create using this connection will use user verification
+    CNCT_specific_data : 7, // Some data, needed for user verification on server
+    CNCT_plugin_name : 8, // Name of plugin, which generated that data
+    CNCT_login : 9, // Same data as isc_dpb_user_name
+    CNCT_plugin_list : 10, // List of plugins, available on client
+    CNCT_client_crypt : 11, // Client encyption level (DISABLED/ENABLED/REQUIRED)
+    WIRE_CRYPT_DISABLED : 0,
+    WIRE_CRYPT_ENABLED : 1,
+    WIRE_CRYPT_REQUIRED : 2,
+};
 
 /****************************/
 /* Common, structural codes */
@@ -161,6 +402,113 @@ const common = {
     isc_info_length                 : 126,
     isc_info_flag_end               : 127,
 };
+
+/*************************************/
+/* Transaction parameter block stuff */
+/*************************************/
+const tpb = {
+    isc_tpb_version1 : 1,
+    isc_tpb_version3 : 3,
+    isc_tpb_consistency : 1,
+    isc_tpb_concurrency : 2,
+    isc_tpb_shared : 3, // < FB21
+    isc_tpb_protected : 4, // < FB21
+    isc_tpb_exclusive : 5, // < FB21
+    isc_tpb_wait : 6,
+    isc_tpb_nowait : 7,
+    isc_tpb_read : 8,
+    isc_tpb_write : 9,
+    isc_tpb_lock_read : 10,
+    isc_tpb_lock_write : 11,
+    isc_tpb_verb_time : 12,
+    isc_tpb_commit_time : 13,
+    isc_tpb_ignore_limbo : 14,
+    isc_tpb_read_committed : 15,
+    isc_tpb_autocommit : 16,
+    isc_tpb_rec_version : 17,
+    isc_tpb_no_rec_version : 18,
+    isc_tpb_restart_requests : 19,
+    isc_tpb_no_auto_undo : 20,
+    isc_tpb_lock_timeout : 21, // >= FB20
+};
+
+const transactionIsolation = {
+    ISOLATION_READ_UNCOMMITTED         : [tpb.isc_tpb_version3, tpb.isc_tpb_write, tpb.isc_tpb_wait, tpb.isc_tpb_read_committed, tpb.isc_tpb_rec_version],
+    ISOLATION_READ_COMMITTED           : [tpb.isc_tpb_version3, tpb.isc_tpb_write, tpb.isc_tpb_wait, tpb.isc_tpb_read_committed, tpb.isc_tpb_no_rec_version],
+    ISOLATION_REPEATABLE_READ          : [tpb.isc_tpb_version3, tpb.isc_tpb_write, tpb.isc_tpb_wait, tpb.isc_tpb_concurrency],
+    ISOLATION_SERIALIZABLE             : [tpb.isc_tpb_version3, tpb.isc_tpb_write, tpb.isc_tpb_wait, tpb.isc_tpb_consistency],
+    ISOLATION_READ_COMMITTED_READ_ONLY : [tpb.isc_tpb_version3, tpb.isc_tpb_read, tpb.isc_tpb_wait, tpb.isc_tpb_read_committed, tpb.isc_tpb_no_rec_version],
+};
+
+/*************************/
+/* SQL information items */
+/*************************/
+const sqlInfo = {
+    isc_info_sql_select : 4,
+    isc_info_sql_bind : 5,
+    isc_info_sql_num_variables : 6,
+    isc_info_sql_describe_vars : 7,
+    isc_info_sql_describe_end : 8,
+    isc_info_sql_sqlda_seq : 9,
+    isc_info_sql_message_seq : 10,
+    isc_info_sql_type : 11,
+    isc_info_sql_sub_type : 12,
+    isc_info_sql_scale : 13,
+    isc_info_sql_length : 14,
+    isc_info_sql_null_ind : 15,
+    isc_info_sql_field : 16,
+    isc_info_sql_relation : 17,
+    isc_info_sql_owner : 18,
+    isc_info_sql_alias : 19,
+    isc_info_sql_sqlda_start : 20,
+    isc_info_sql_stmt_type : 21,
+    isc_info_sql_get_plan : 22,
+    isc_info_sql_records : 23,
+    isc_info_sql_batch_fetch : 24,
+    isc_info_sql_relation_alias : 25, // >: 2.0
+    isc_info_sql_explain_plan : 26, // >= 3.0
+};
+
+const statementInfo = {
+    isc_info_sql_stmt_select : 1,
+    isc_info_sql_stmt_insert : 2,
+    isc_info_sql_stmt_update : 3,
+    isc_info_sql_stmt_delete : 4,
+    isc_info_sql_stmt_ddl : 5,
+    isc_info_sql_stmt_get_segment : 6,
+    isc_info_sql_stmt_put_segment : 7,
+    isc_info_sql_stmt_exec_procedure : 8,
+    isc_info_sql_stmt_start_trans : 9,
+    isc_info_sql_stmt_commit : 10,
+    isc_info_sql_stmt_rollback : 11,
+    isc_info_sql_stmt_select_for_upd : 12,
+    isc_info_sql_stmt_set_generator : 13,
+    isc_info_sql_stmt_savepoint : 14,
+};
+
+const DESCRIBE = [
+    sqlInfo.isc_info_sql_stmt_type,
+    sqlInfo.isc_info_sql_select,
+    sqlInfo.isc_info_sql_describe_vars,
+    sqlInfo.isc_info_sql_sqlda_seq,
+    sqlInfo.isc_info_sql_type,
+    sqlInfo.isc_info_sql_sub_type,
+    sqlInfo.isc_info_sql_scale,
+    sqlInfo.isc_info_sql_length,
+    sqlInfo.isc_info_sql_field,
+    sqlInfo.isc_info_sql_relation,
+    //isc_info_sql_owner,
+    sqlInfo.isc_info_sql_alias,
+    sqlInfo.isc_info_sql_describe_end,
+    sqlInfo.isc_info_sql_bind,
+    sqlInfo.isc_info_sql_describe_vars,
+    sqlInfo.isc_info_sql_sqlda_seq,
+    sqlInfo.isc_info_sql_type,
+    sqlInfo.isc_info_sql_sub_type,
+    sqlInfo.isc_info_sql_scale,
+    sqlInfo.isc_info_sql_length,
+    sqlInfo.isc_info_sql_describe_end
+];
 
 /***********************/
 /*   ISC Services      */
@@ -314,7 +662,6 @@ const serviceRestore = {
     isc_spb_res_use_all_space : 0x4000,
 };
 
-
 /* · Repair Service ·*/
 const serviceRepair = {
     isc_spb_rpr_commit_trans : 15,
@@ -386,10 +733,24 @@ const serviceTrace = {
 };
 
 module.exports = Object.freeze({
+    ...acceptType,
+    ...authPlugin,
+    ...authOptions,
     ...blr,
+    ...blobType,
+    ...buffer,
+    ...cnct,
     ...common,
+    ...connect,
+    ...defaultOptions,
+    DESCRIBE,
     ...dpb,
+    ...dsql,
+    ...int,
     ...iscAction,
+    ...iscError,
+    ...op,
+    ...protocol,
     ...service,
     ...serviceBackup,
     ...serviceInfo,
@@ -399,5 +760,11 @@ module.exports = Object.freeze({
     ...serviceSecurity,
     ...serviceStatistics,
     ...serviceTrace,
-    ...spb
+    ...sqlInfo,
+    ...sqlType,
+    ...spb,
+    ...statementInfo,
+    SUPPORTED_PROTOCOL,
+    ...tpb,
+    ...transactionIsolation,
 });
