+++ lib/wire/serialize.js
