+++ lib/wire/service.js
@@ -1,7 +1,7 @@
 const Events = require("events");
 const stream = require("stream");
 const Const = require('./const');
-const {BlrReader} = require("../serialize");
+const {BlrReader} = require("./serialize");
 const {doError} = require("../callback");
 
 /***************************************
